+++ src/invoices.js
@@ -5,6 +5,7 @@ class Invoices {
   constructor (wrapper) {
     this.wrapper = wrapper;
   }
+
   /**
    * Creates a new valid invoice (CFDI).
    * @param {Object} data
@@ -13,6 +14,7 @@ class Invoices {
   create (data) {
     return this.wrapper.createInvoice(data);
   }
+
   /**
    * Gets a paginated list of invoices created by your organization
    * @param {[Object]} params - Search parameters
@@ -21,6 +23,7 @@ class Invoices {
   list (params) {
     return this.wrapper.listInvoices(params);
   }
+
   /**
    * Gets a single invoice object
    * @param {string} id
@@ -29,6 +32,7 @@ class Invoices {
   retrieve (id) {
     return this.wrapper.retrieveInvoice(id);
   }
+
   /**
    * Cancels an invoice. The invoice will not be valid anymore and will change its status to canceled.
    * @param {string} id
@@ -37,6 +41,7 @@ class Invoices {
   cancel (id) {
     return this.wrapper.cancelInvoice(id);
   }
+
   /**
    * Sends the invoice to the customer's email
    * @param {String} id Invoice Id
@@ -47,6 +52,7 @@ class Invoices {
   sendByEmail (id, data) {
     return this.wrapper.sendInvoiceByEmail(id, data);
   }
+
   /**
    * Downloads the specified invoice in PDF format
    * @param {string} id Invoice Id
@@ -55,6 +61,7 @@ class Invoices {
   downloadPdf (id) {
     return this.wrapper.downloadPdf(id);
   }
+
   /**
    * Downloads the specified invoice in XML format
    * @param {string} id Invoice Id
@@ -63,6 +70,7 @@ class Invoices {
   downloadXml (id) {
     return this.wrapper.downloadXml(id);
   }
+
   /**
    * Downloads the specified invoice in a ZIP package containing both PDF and XML files
    * @param {string} id Invoice Id